+++ packages/ui/CHANGELOG.md
@@ -7,12 +7,22 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Changed
 
 - `ui/ToggleItem` to to be slottable
+### Added
+
+- `ui/Touchable` support to fire `onTap` when a `click` event occurs
+
+### Changed
+
+- `ui/Touchable` custom events `onDown`, `onUp`, `onMove`, and `onTap` to use the event name as the `type` rather than the shorter name (e.g. `onTap` rather than `tap`)
+- `ui/Toggleable` to forward events on `activate` and `deactivate` instead of firing toggled payload. Use `toggle` to handle toggled payload from the event
+
+## [2.0.0-beta.2] - 2018-05-07
 
 ### Fixed
 
 - `ui/Marquee` to always marquee when `marqueeOn` is set to `'render'`
 - `ui/Item` to use its natural width rather than imposing a 100% width allowing inline Items to be the correct width
-- `ui/MarqueeDecorator` to correctly reset animation when children updates
+- `ui/Marquee.MarqueeDecorator` to correctly reset animation when `children` updates
 
 ## [2.0.0-beta.1] - 2018-04-29